+++ backend/canisters/online_users/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Track user online minutes per month ([#7547](https://github.com/open-chat-labs/open-chat/pull/7547))
+
 ### Changed
 
 - Introduce `StableMemoryMap` trait to simplify storing in stable memory ([#7176](https://github.com/open-chat-labs/open-chat/pull/7176))